+++ kroto-plus-compiler/src/main/kotlin/com/github/marcoferrer/krotoplus/generators/ProtoTypeBuilderGenerator.kt
@@ -37,7 +37,7 @@ class ProtoTypeBuilderGenerator(
                 }
     }
 
-    private suspend fun buildFileSpec(protoFile: ProtoFile){
+    private fun buildFileSpec(protoFile: ProtoFile){
         val filename = "${protoFile.javaOuterClassname}Builders"
         val fileSpecBuilder = FileSpec.builder(protoFile.outputPackage(),filename)
                 .addComment(AutoGenerationDisclaimer)
@@ -49,6 +49,9 @@ class ProtoTypeBuilderGenerator(
         val typeSpecBuilder = TypeSpec.objectBuilder(filename)
                 .addAnnotation(protoFile.getGeneratedAnnotationSpec())
 
+        val builderFunSpecs = mutableListOf<FunSpec>()
+        val copyFunSpecs = mutableListOf<FunSpec>()
+
         protoFile.types()
                 .asSequence()
                 .filterNot { it is EnumType }
@@ -56,25 +59,43 @@ class ProtoTypeBuilderGenerator(
                 .forEach { protoType ->
 
                     val typeClassName = protoType.toClassName(protoFile)
+                    val builderTypeName = LambdaTypeName.get(
+                            receiver = typeClassName.nestedClass("Builder"),
+                            returnType = UNIT)
 
+                    // Make optional with generator option?
                     FunSpec.builder(protoType.simpleName())
                             .addModifiers(KModifier.INLINE)
-                            .addParameter("block",LambdaTypeName.get(
-                                receiver = typeClassName.nestedClass("Builder"),
-                                returnType = UNIT))
+                            .addParameter("block", builderTypeName)
                             .addStatement("return %T.newBuilder().apply(block).build()", typeClassName)
                             .also {
-                                typeSpecBuilder.addFunction(it.build())
+                                builderFunSpecs.add(it.build())
+                            }
+
+                    // Make optional with generator option?
+                    FunSpec.builder("copy")
+                            .receiver(typeClassName)
+                            .addModifiers(KModifier.INLINE)
+                            .addParameter("block",builderTypeName)
+                            .addStatement("return this.toBuilder().apply(block).build()")
+                            .also {
+                                copyFunSpecs.add(it.build())
                             }
                 }
 
-        typeSpecBuilder.build()
-                .takeIf { it.funSpecs.isNotEmpty() }
-                ?.let { typeSpec ->
+        if(builderFunSpecs.isNotEmpty()){
+            typeSpecBuilder.addFunctions(builderFunSpecs)
+            fileSpecBuilder.addType(typeSpecBuilder.build())
+        }
 
-                    val fileSpec = fileSpecBuilder.addType(typeSpec).build()
-                    resultChannel.send(GeneratorResult(fileSpec, defaultOutputDir))
-                }
+        for(funSpec in copyFunSpecs)
+            fileSpecBuilder.addFunction(funSpec)
+
+        // This is a redundant check on any funSpec collection. This is done in case in the future
+        // we decide to omit certain proto types from either collection
+        if(builderFunSpecs.isNotEmpty() || copyFunSpecs.isNotEmpty())
+            resultChannel.offer(GeneratorResult(fileSpecBuilder.build(), defaultOutputDir))
+
     }