+++ CHANGELOG.md
@@ -1,5 +1,26 @@
-## Version 0.1.0
+## Version 0.1.2
+_2018-04-23_
 
+* New: Updated Kotlin runtime to ```1.2.40```
+
+#### Compiler
+* Fix: Generated pom no longer includes dependencies that have been embedded 
+
+## Version 0.1.1
+_2018-04-10_
+
+#### Coroutines
+* New: Added ```data``` modifier to ```ClientBidiCallChannel``` to support destructuring declarations
+* New: Added ```ServerBidiCallChannel<ReqT,RespT>``` for future server side coroutine support 
+* Fix: Renamed ```RpcBidiChannel<ReqT,RespT>``` to ```ClientBidiCallChannel<ReqT,RespT>``` 
+
+#### Gradle Plugin 
+* New: Published to gradle plugin portal
+
+#### Compiler
+* New: [GH-1](https://github.com/marcoferrer/kroto-plus/issues/1) Added copy extension for proto types.
+
+## Version 0.1.0
 _2018-03-26_
 
  * Initial release.